+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - Custom Roles can now have any parameter type they would like, fixing #925
 - Install-Lab imports the VMs RDP certificates and Remove-Lab removes them to enable seamless Connect-LabVm
 - Relaxed Azure password policy as special characters are not mandatory.
+- Windows Admin Center implemented as proper Role to enable SkipDeployment parameter
   
 ### Bug Fixes
 - Get-LabInternetFile did not work on Azure when the Uri did not contain a file name like 'https://go.microsoft.com/fwlink/?Linkid=85215'.